rename: 테스트 코드 도메인별 패키징
