+++ src/test/java/domain/ladder/LadderTest.java
@@ -1,7 +1,5 @@
-package domain;
+package domain.ladder;
 
-import domain.ladder.Ladder;
-import domain.ladder.Line;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
